chore: move modules out of folders
