--- packages/app-runtime/src/modules/appSync/index.ts
@@ -1,1 +1,0 @@
-export * from "./AppSyncModule";
