+++ packages/app-runtime/src/modules/index.ts
@@ -1,5 +1,5 @@
 export * from "./appEvents";
 export * from "./AppRuntimeModule";
-export * from "./appSync";
-export * from "./pushNotifications";
+export * from "./AppSyncModule";
+export * from "./PushNotificationModule";
 export * from "./runtimeEvents";
