--- packages/app-runtime/src/modules/pushNotifications/IBackboneEventContent.ts
@@ -1,11 +1,0 @@
-export enum BackboneEventName {
-    DatawalletModificationsCreated = "DatawalletModificationsCreated",
-    ExternalEventCreated = "ExternalEventCreated"
-}
-
-export interface IBackboneEventContent {
-    devicePushIdentifier: string;
-    eventName: BackboneEventName;
-    sentAt: string;
-    payload: any;
-}
