--- packages/app-runtime/src/modules/pushNotifications/index.ts
@@ -1,2 +1,0 @@
-export * from "./IBackboneEventContent";
-export * from "./PushNotificationModule";
